requirements.txt
feedparser==6.0.10
requests==2.31.0
beautifulsoup4==4.12.2
openai==0.28.1
notion-client==2.2.1
python-dotenv==1.0.0
lxml==4.9.3 